+++ Client/package.json
@@ -12,6 +12,7 @@
     "@expo/ngrok": "^4.1.3",
     "@expo/webpack-config": "^19.0.0",
     "@react-native-async-storage/async-storage": "^1.19.5",
+    "@react-native-firebase/app": "^18.7.3",
     "@react-native-picker/picker": "^2.6.1",
     "@react-navigation/bottom-tabs": "^6.5.11",
     "@react-navigation/native": "^6.1.9",
@@ -20,6 +21,8 @@
     "@rneui/themed": "^4.0.0-rc.8",
     "expo": "~49.0.15",
     "expo-cli": "^6.3.10",
+    "expo-constants": "^14.4.2",
+    "expo-dev-client": "~2.4.12",
     "expo-device": "^5.6.0",
     "expo-linear-gradient": "~12.3.0",
     "expo-notifications": "^0.23.0",
